Migrate to circleci 2.0 and add requirements.txt
