+++ requirements.txt
@@ -1,0 +1,5 @@
+pytest
+pytest-cov
+python-dateutil
+aniso8601
+JPype1
